search-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.9
pydantic==2.11.7
redis==6.4.0
kafka-python==2.2.15
httpx==0.28.1
mcp==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
grpcio==1.74.0
grpcio-tools==1.74.0
protobuf==6.32.0
